[REFACTOR]make focus handler function as part of useFormSection hook so that it can be reused in different form components
